requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
python-multipart
pydantic
langchain-text-splitters
motor
python-dotenv
pymongo
certifi
langchain 
faiss-cpu
google-genai
numpy
sentence-transformers
pinecone